# Trace all ExprssionEvaluator::evaluate